Cambios en migraciones y creación de seeders

+++ database/migrations/2020_05_04_014256_create_coupons_table.php
@@ -17,7 +17,7 @@ class CreateCouponsTable extends Migration
             $table->bigIncrements('id');
             $table->string('codigo')->unique();
             $table->string('coupon_type');
-            $table->integer('coupon_id');
+            $table->unsignedBigInteger('coupon_id');
             $table->softDeletes();
             $table->timestamps();
         });
